ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
pydantic==2.8.2
google-genai==1.46.0
torch
transformers
sentencepiece
